Moved basic postprocessor implementation to a standalone file, fixed imports.
