+++ src/ast/ast_reader.ts
@@ -2,7 +2,7 @@ import { ASTNode, ASTNodeConstructor } from "./ast_node";
 import { SourceUnit } from "./implementation/meta/source_unit";
 import { LegacyConfiguration } from "./legacy";
 import { ModernConfiguration } from "./modern";
-import { ASTPostprocessor } from "./postprocessing/postprocessor";
+import { DefaultPostprocessorMapping } from "./postprocessing";
 import { sequence } from "./utils";
 
 export interface ASTNodeProcessor<T extends ASTNode> {
@@ -11,6 +11,12 @@ export interface ASTNodeProcessor<T extends ASTNode> {
         config: ASTReaderConfiguration,
         raw: any
     ): ConstructorParameters<ASTNodeConstructor<T>>;
+}
+
+export interface ASTNodePostprocessor {
+    readonly priority: number;
+
+    process(node: ASTNode, context: ASTContext, sources?: Map<string, string>): void;
 }
 
 export interface ASTReadingRule {
@@ -112,6 +118,59 @@ export class ASTContext {
 
     contains(node: ASTNode): boolean {
         return this.locate(node.id) === node;
+    }
+}
+
+export class ASTPostprocessor {
+    mapping: Map<ASTNodeConstructor<ASTNode>, ASTNodePostprocessor[]>;
+
+    constructor(mapping = DefaultPostprocessorMapping) {
+        this.mapping = mapping;
+    }
+
+    getPostprocessorsForNode(node: ASTNode): ASTNodePostprocessor[] | undefined {
+        return this.mapping.get(node.constructor as ASTNodeConstructor<ASTNode>);
+    }
+
+    processNode(node: ASTNode, context: ASTContext, sources?: Map<string, string>): void {
+        const postprocessors = this.getPostprocessorsForNode(node);
+
+        if (postprocessors) {
+            for (const postprocessor of postprocessors) {
+                postprocessor.process(node, context, sources);
+            }
+        }
+    }
+
+    processContext(context: ASTContext, sources?: Map<string, string>): void {
+        const groupsByPriority = new Map<number, ASTNode[]>();
+
+        for (const node of context.nodes) {
+            const postprocessors = this.getPostprocessorsForNode(node);
+
+            if (postprocessors) {
+                for (const postprocessor of postprocessors) {
+                    const priority = postprocessor.priority;
+                    const group = groupsByPriority.get(priority);
+
+                    if (group) {
+                        group.push(node);
+                    } else {
+                        groupsByPriority.set(priority, [node]);
+                    }
+                }
+            }
+        }
+
+        const groups = Array.from(groupsByPriority)
+            .sort((a, b) => a[0] - b[0])
+            .map((entry) => entry[1]);
+
+        for (const nodes of groups) {
+            for (const node of nodes) {
+                this.processNode(node, context, sources);
+            }
+        }
     }
 }
