+++ src/ast/postprocessing/index.ts
@@ -1,4 +1,4 @@
-export * from "./mapping";
 export * from "./builtin_referenced_declaration_normalizer";
+export * from "./mapping";
+export * from "./node_postprocessor";
 export * from "./structured_documentation_reconstruction";
-export * from "./postprocessor";
