+++ src/ast/postprocessing/mapping.ts
@@ -1,4 +1,5 @@
 import { ASTNode, ASTNodeConstructor } from "../ast_node";
+import { ASTNodePostprocessor } from "../ast_reader";
 import { ContractDefinition } from "../implementation/declaration/contract_definition";
 import { EventDefinition } from "../implementation/declaration/event_definition";
 import { FunctionDefinition } from "../implementation/declaration/function_definition";
@@ -9,7 +10,6 @@ import { MemberAccess } from "../implementation/expression/member_access";
 import { IdentifierPath } from "../implementation/meta/identifier_path";
 import { UserDefinedTypeName } from "../implementation/type/user_defined_type_name";
 import { BuiltinReferencedDeclarationNormalizer } from "./builtin_referenced_declaration_normalizer";
-import { ASTNodePostprocessor } from "./postprocessor";
 import { StructuredDocumentationReconstructingPostprocessor } from "./structured_documentation_reconstruction";
 
 const reconstructor = new StructuredDocumentationReconstructingPostprocessor(1000);
